% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{annotation.mq}
\alias{annotation.mq}
\title{Example annotation file for MaxQuant}
\format{
data.frame
}
\description{
data.frame mapping MaxQuant run names to the corresponding bioreplicates and
conditions. Used as input to preprocessing function, converting data into
MSstats format.
}
\examples{
data(annotation.mq)
head(annotation.mq)

}
