% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{annotation.pd}
\alias{annotation.pd}
\title{Example annotation file for PD}
\format{
data.frame
}
\description{
data.frame mapping PD run names to the corresponding bioreplicates and
conditions. Used as input to preprocessing function, converting data into
MSstats format.
}
\examples{
data(annotation.pd)
head(annotation.pd)

}
