% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{dia_skyline_model}
\alias{dia_skyline_model}
\title{Example of Sklyine DDA dataset modeled using MSstats \code{groupComparison} function.}
\format{
list
}
\description{
Data includes one list with two data.tables named \code{ComparisonResult} and
\code{ModelQC} and another list of model details named \code{FittedModel}.
\code{ComparisonResult} shows an overview of all proteins modeled in the system.
\code{ModelQC} provides a report on the quality control checks of each protein
in the dataset.
}
\examples{
data(dia_skyline_model)
head(dia_skyline_model)

}
