% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{dia_skyline_summarized}
\alias{dia_skyline_summarized}
\title{Example of Sklyine DDA dataset processed using MSstats summarization function.}
\format{
list
}
\usage{
data(dia_skyline_summarized)
}
\description{
Data includes one list with two data.tables named \code{FeatureLevelData} and
\code{ProteinLevelData} and a string value \code{SummaryMethod}. \code{FeatureLevelData}
shows the unsummarized feature level data. \code{ProteinLevelData} shows the data
summarized up to the protein level and is used for modeling the data.
}
\examples{
data("dia_skyline_summarized")
head(dia_skyline_summarized)

}
\keyword{datasets}
