% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{evidence}
\alias{evidence}
\title{Example evidence file for MaxQuant}
\format{
data.frame
}
\description{
data.frame containing output of MaxQuant. Used in examples.
}
\examples{
data(evidence)
head(evidence)

}
