% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{example_dia_skyline}
\alias{example_dia_skyline}
\title{Example of input Sklyine DDA dataset.}
\format{
data.frame
}
\usage{
data(example_dia_skyline)
}
\description{
Used as input data to MSstats workflow. Data includes one data.table which
is the output of Skyline.
}
\details{
The raw data (input data for MSstats) is required to contain
variable of ProteinName, PeptideSequence, PrecursorCharge, FragmentIon,
ProductCharge, IsotopeLabelType, Condition, BioReplicate, Run, Intensity.
The variable names should be fixed. If the information of one or more columns
is not available for the original raw data, please retain the column
variables and type in fixed value. For example, the original raw data does
not contain the information of PrecursorCharge and ProductCharge, we retain
the column PrecursorCharge and ProductCharge and then type in NA for all
transitions in RawData. Variable Intensity is required to be original signal
without any log transformation and can be specified as the peak of height
or the peak of  area under curve.
}
\examples{
data("example_dia_skyline")
head(example_dia_skyline)

}
\keyword{datasets}
