% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{example_skyline_annotation}
\alias{example_skyline_annotation}
\title{Example annotation file}
\format{
data.frame
}
\description{
data.frame mapping Skyline run names to the corresponding bioreplicates and
conditions. Used as input to preprocessing function, converting data into
MSstats format.
}
\examples{
data(example_skyline_annotation)
head(example_skyline_annotation)

}
