% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmodel-server-visualization.R
\name{get_download_plot_filename}
\alias{get_download_plot_filename}
\title{Get filename for plot download based on plot type}
\usage{
get_download_plot_filename(plot_type)
}
\arguments{
\item{plot_type}{the current plot type string}
}
\value{
filename string ending in .zip
}
\description{
Get filename for plot download based on plot type
}
