% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-home-ui.R
\name{homeUI}
\alias{homeUI}
\title{Home UI module for home page.}
\usage{
homeUI(id)
}
\arguments{
\item{id}{namespace prefix for the module}
}
\value{
This function returns nothing, as it sets up the Home UI
}
\description{
This function generates the home user interface for MSstatsShiny, a web tool
for the statistical analysis of quantitative proteomic data built around the
R packages MSstats, MSstatsTMT, and MSstatsPTM.
}
