% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-mssstats-help-ui.R
\name{msstatsHelpUI}
\alias{msstatsHelpUI}
\title{Help MSStats UI module for msstats help page.}
\usage{
msstatsHelpUI(id)
}
\arguments{
\item{id}{namespace prefix for the module}
}
\value{
This function returns nothing, as it sets up the MSStats Help UI
}
\description{
This module shows the msstats help page for general documentation
}
