% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-statstmt-help-ui.R
\name{msstatsTmtHelpUI}
\alias{msstatsTmtHelpUI}
\title{Help MSStats UI module for msstatstmt help page.}
\usage{
msstatsTmtHelpUI(id)
}
\arguments{
\item{id}{namespace prefix for the module}
}
\value{
This function returns nothing, as it sets up the MSStatstmts Help UI
}
\description{
This module shows the msstats help page for general documentation
}
\examples{
\dontrun{
library(shiny)
ui <- fluidPage(
  msstatsTmtHelpUI("tmt")
)
server <- function(input, output, session) {
  # Server logic for msstatsTmtHelpUI
}
shinyApp(ui, server)
}
}
