% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-visualize-network-ui.R
\name{networkUI}
\alias{networkUI}
\title{Network UI module for visualizing protein regulatory networks from INDRA.}
\usage{
networkUI(id)
}
\arguments{
\item{id}{namespace prefix for the module}
}
\description{
Network UI module for visualizing protein regulatory networks from INDRA.
}
