% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-expdes-server.R
\name{plot_tpr_power_curve}
\alias{plot_tpr_power_curve}
\title{Plot TPR power curves for dose response proteomics experiments}
\usage{
plot_tpr_power_curve(simulation_results)
}
\arguments{
\item{simulation_results}{data.frame from run_tpr_simulation}
}
\value{
plotly object with two panels (Strong and Weak interaction)
}
\description{
Plot TPR power curves for dose response proteomics experiments
}
