% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{proteinGroups}
\alias{proteinGroups}
\title{Example ProteinGroups file for MaxQuant}
\format{
data.frame
}
\description{
data.frame containing output of ProteinGroups MaxQuant file. Used in examples.
}
\examples{
data(proteinGroups)
head(proteinGroups)

}
