% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_calculations.R
\name{QC_check}
\alias{QC_check}
\title{Quick QC value check}
\usage{
QC_check(input)
}
\arguments{
\item{input}{options for data processing input by the user}
}
\value{
string
}
\description{
Quick QC value check for LF vs TMT
}
\examples{
input = list(null=TRUE)
QC_check(input)
}
