% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-qc-ui.R
\name{qcUI}
\alias{qcUI}
\title{QC UI module for data processing UI.}
\usage{
qcUI(id)
}
\arguments{
\item{id}{namespace prefix for the module}
}
\value{
This function returns nothing, as it sets up the QC UI
}
\description{
This function sets up the QC UI where it consists of several,
options for users to process data based on previously selected fragments.
}
