% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_calculations.R
\name{radioTooltip}
\alias{radioTooltip}
\title{Custom function to create radio tool tips}
\usage{
radioTooltip(
  id,
  choice,
  title,
  placement = "bottom",
  trigger = "hover",
  options = NULL
)
}
\arguments{
\item{id}{input id}

\item{choice}{user selection}

\item{title}{title of object}

\item{placement}{where should tooltip be shown}

\item{trigger}{how should prompt be shown}

\item{options}{additional options to pass to function}
}
\value{
HTML object
}
\description{
Used in UI files to create HTML vizualizations
}
\examples{
radioTooltip("testid", "test_choice", "test_title")
}
