% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{raw.om}
\alias{raw.om}
\title{Example output file Spectromine}
\format{
data.frame
}
\description{
data.frame containing output of Spectromine. Used in examples.
}
\examples{
data(raw.om)
head(raw.om)

}
