% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-statmodel-ui.R
\name{statmodelUI}
\alias{statmodelUI}
\title{Statmodel UI module for statistical inference UI.}
\usage{
statmodelUI(id)
}
\arguments{
\item{id}{namespace prefix for the module}
}
\value{
This function returns nothing, as it sets up the Statmodel UI
}
\description{
This function sets up the Statmodel UI where it consists of several,
options for users to select and upload files.
}
