% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDoc.R
\docType{data}
\name{tmt_pd_model}
\alias{tmt_pd_model}
\title{Example of TMT dataset modeled using MSstatsTMT \code{groupComparisonTMT} function.}
\format{
list
}
\description{
Data includes one list with two data.tables named \code{ComparisonResult} and
\code{ModelQC} and another list of model details named \code{FittedModel}.
\code{ComparisonResult} shows an overview of all proteins modeled in the system.
\code{ModelQC} provides a report on the quality control checks of each protein
in the dataset.
}
\examples{
data(tmt_pd_model)
head(tmt_pd_model)

}
