% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{uiObject}
\alias{uiObject}
\title{UI function for the MSstatsShiny app}
\usage{
uiObject()
}
\value{
UI object for shinyUI
}
\description{
This functions generates the UI object for MSstatsShiny app. Responsible
for generating 5 nav pages homepage, data upload page, data processing page, statistical
inference and future experiments.
}
\examples{
NA

}
