% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-visualize-network-server.R
\name{visualizeNetworkServer}
\alias{visualizeNetworkServer}
\title{Server logic for network visualization module}
\usage{
visualizeNetworkServer(input, output, session, parent_session, dataComparison)
}
\arguments{
\item{input}{Shiny input object}

\item{output}{Shiny output object}

\item{session}{Shiny session object}

\item{parent_session}{Parent Shiny session}

\item{dataComparison}{Reactive expression containing comparison data}
}
\value{
None (side effects only)
}
\description{
Server logic for network visualization module
}
