% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_calculations.R
\name{xy_str}
\alias{xy_str}
\title{Simple function to return coordinates}
\usage{
xy_str(e)
}
\arguments{
\item{e}{input function provided by user}
}
\value{
Character with x and y coordinates
}
\description{
Used in experimental design to create vizualization
}
\examples{
xy_str(list(x=5.0,y=2.0))
}
