% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statmodel-server-visualization.R
\name{zip_and_copy_plot}
\alias{zip_and_copy_plot}
\title{Zip PDF files and copy to download destination}
\usage{
zip_and_copy_plot(pdf_files, dest_file)
}
\arguments{
\item{pdf_files}{character vector of PDF file paths}

\item{dest_file}{destination file path for the download}
}
\value{
TRUE if successful, FALSE otherwise
}
\description{
Zip PDF files and copy to download destination
}
