% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_calculations.R
\name{apply_adj}
\alias{apply_adj}
\title{Main PTM adjustment function}
\usage{
apply_adj(ptm_model, protein_model)
}
\arguments{
\item{ptm_model}{output of MSstats modeling function modeling PTMs}

\item{protein_model}{output of MSstats modeling function modeling unmodified proteins}
}
\value{
list of PTM modeling results
}
\description{
Main PTM function to model MSstatsShiny data.
}
\examples{
#apply_adj(NULL)

}
