% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-expdes-ui.R
\name{expdesUI}
\alias{expdesUI}
\title{Expdes UI module for future experiments UI.}
\usage{
expdesUI(id)
}
\arguments{
\item{id}{namespace prefix for the module}
}
\value{
This function returns nothing, as it sets up the Expdes UI
}
\description{
This function sets up the Expdes UI where it consists of several,
options for users to select and generate plots.
}
\examples{
\dontrun{
library(shiny)
library(shinyjs)

# Define UI for app that incorporates Expdes UI
ui <- fluidPage(
  expdesUI("expdes")
)

# Define server logic to interact with Expdes UI
server <- function(input, output, session) {
  # Placeholder for server-side logic of Expdes module
}

# Run the application
shinyApp(ui = ui, server = server)
}
}
