% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-help-ui.R
\name{helpUI}
\alias{helpUI}
\title{Help UI module for help page.}
\usage{
helpUI(id)
}
\arguments{
\item{id}{namespace prefix for the module}
}
\value{
This function returns nothing, as it sets up the Help UI
}
\description{
This module shows the help page for general documentation
}
\examples{
NA

}
