% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_MSstatsShiny.R
\name{launch_MSstatsShiny}
\alias{launch_MSstatsShiny}
\title{Run MSstatsShiny Application}
\usage{
launch_MSstatsShiny(
  launch_app = TRUE,
  port = getOption("shiny.port"),
  host = getOption("shiny.host", "127.0.0.1"),
  testMode = FALSE
)
}
\arguments{
\item{launch_app}{One of TRUE or FALSE indicating whether or not to run
application. Default is TRUE.}

\item{port}{(optional) Specify port the application should list to.}

\item{host}{(optional) The IPv4 address that the application should listen on.}

\item{testMode}{One of TRUE or FALSE indicating whether or not to run the
application in test mode. Default is FALSE.}
}
\value{
Running Shiny Application
}
\description{
Main function to run MSstatsShiny. All other functions in this package are
run automatically.
}
\examples{
\dontrun{
## To run app set launch_app=TRUE
launch_MSstatsShiny(launch_app=FALSE,testMode=FALSE)
}
}
