% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_MSstatsShiny.R
\name{launch_MSstatsShiny}
\alias{launch_MSstatsShiny}
\title{Run MSstatsShiny Application}
\usage{
launch_MSstatsShiny(launch_app = TRUE)
}
\arguments{
\item{launch_app}{One of TRUE or FALSE indicating wheter or not to run
application. Default is TRUE.}
}
\value{
Running Shiny Application
}
\description{
Main function to run MSstatsShiny. All other functions in this package are
run automatically.
}
\examples{

launch_MSstatsShiny(launch_app=FALSE)

}
