% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_calculations.R
\name{lf_model}
\alias{lf_model}
\title{Main LF modeling function for MSstatsShiny application}
\usage{
lf_model(data, contrast.matrix, busy_indicator = TRUE)
}
\value{
list of LF modeling results
}
\description{
Main LF function to model MSstatsShiny data.
}
\examples{
#lf_model(NULL)

}
