% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_calculations.R
\name{lf_summarization_loop}
\alias{lf_summarization_loop}
\title{Main LF calculation summarization function for MSstatsShiny application}
\usage{
lf_summarization_loop(data, qc_input, loadpage_input, busy_indicator = TRUE)
}
\arguments{
\item{data}{Data converted into MSstats format.}

\item{qc_input}{options for data processing input by the user from data processing page.}

\item{loadpage_input}{options for data processing input by the user from data upload page.}

\item{busy_indicator}{Boolean indicator indicating whether or not to display
shiny waiting indicator.}
}
\value{
list of LF Summarization results
}
\description{
Main LF function to calculate MSstatsShiny results.
}
