% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_calculations.R
\name{lf_summarization_loop}
\alias{lf_summarization_loop}
\title{Main LF calculation summarization function for MSstatsShiny application}
\usage{
lf_summarization_loop(data, input, busy_indicator = TRUE)
}
\arguments{
\item{data}{Data converted into MSstats format.}

\item{input}{options for data processing input by the user}

\item{busy_indicator}{Boolean indicator indicating whether or not to display
shiny waiting indicator.}
}
\value{
list of LF Summarization results
}
\description{
Main LF function to calculate MSstatsShiny results.
}
\examples{
data("example_dia_skyline")
data("example_skyline_annotation")
testdata = MSstats::SkylinetoMSstatsFormat(example_dia_skyline,
                                            annotation = example_skyline_annotation,
                                            filter_with_Qvalue = TRUE, 
                                            qvalue_cutoff = 0.01, 
                                            fewMeasurements="remove", 
                                            removeProtein_with1Feature = TRUE,
                                            use_log_file = FALSE)

## Source app functionality
input = list()
input$norm = "equalizeMedians"
input$log = 2
input$names = NULL
input$features_used	= "all"
code_n_feat=3
input$censInt = "NA"
input$features_used	= "all"
input$MBi = TRUE
input$remove50 = FALSE
input$maxQC = 0.999
input$null = FALSE
input$null1 = FALSE
input$DDA_DIA = "LF"

lf_summarization_loop(testdata, input, busy_indicator=FALSE)

}
