% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-loadpage-server.R
\name{loadpageServer}
\alias{loadpageServer}
\title{Loadpage Server module for data selection and upload server.}
\usage{
loadpageServer(id, parent_session)
}
\arguments{
\item{id}{namespace prefix for the module}

\item{parent_session}{session of the main calling module}
}
\value{
input object with user selected options
}
\description{
This function sets up the loadpage server where it consists of several,
options for users to select and upload files.
}
