% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-loadpage-ui.R
\name{loadpageUI}
\alias{loadpageUI}
\title{Loadpage UI module for data selection and upload UI.}
\usage{
loadpageUI(id)
}
\arguments{
\item{id}{namespace prefix for the module}
}
\value{
This function returns nothing, as it sets up the loadpage UI
}
\description{
This function sets up the loadpage UI where it consists of several,
options for users to select and upload files.
}
\examples{
\dontrun{
library(shiny)

# Define UI for app that incorporates Help UI
ui <- fluidPage(
  loadpageUI("loadpage")
)

# Define server logic to interact with Help UI
server <- function(input, output, session) {
  # Placeholder for server-side logic of Help module
}

# Run the application
shinyApp(ui = ui, server = server)
}
}
