% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_calculations.R
\name{QC_check}
\alias{QC_check}
\title{Quick QC value check}
\usage{
QC_check(qc_input, loadpage_input)
}
\arguments{
\item{qc_input}{options for data processing input by the user from data processing page.}

\item{loadpage_input}{options for data processing input by the user from data upload page.}
}
\value{
string
}
\description{
Quick QC value check for LF vs TMT
}
\examples{
qc_input = list(null=TRUE)
loadpage_input = list(null=TRUE)
QC_check(qc_input,loadpage_input)
}
