% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-qc-server.R
\name{qcServer}
\alias{qcServer}
\title{QC Server module for data processing}
\usage{
qcServer(input, output, session, parent_session, loadpage_input, get_data)
}
\arguments{
\item{input}{input object to capture different ui element values}

\item{output}{to render and create elements}

\item{session}{session current module}

\item{parent_session}{session of the main calling module}

\item{loadpage_input}{input object from loadpage UI}
}
\value{
input object with user selected options
}
\description{
This function sets up the QC server to process data based on user
selected inputs
}
