% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server function for the MSstatsShiny app}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{shiny server input}

\item{output}{shiny server output}

\item{session}{session object for shiny to connect to}
}
\value{
Server object for shinyUI
}
\description{
This functions generates the Server object for MSstatsShiny app.
}
\examples{
NA

}
