% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-statmodel-server.R
\name{statmodelServer}
\alias{statmodelServer}
\title{Statmodel Server module for stat inference}
\usage{
statmodelServer(
  id,
  parent_session,
  loadpage_input,
  qc_input,
  get_data,
  preprocess_data,
  app_template = reactive(TEMPLATES$default),
  turnover_ratios = reactive(NULL),
  condition_metadata = reactive(NULL)
)
}
\arguments{
\item{id}{namespace prefix for the module}

\item{parent_session}{session of the main calling module}

\item{loadpage_input}{input object from loadpage UI}

\item{qc_input}{input object from QC UI}

\item{get_data}{stored function that returns the data from loadpage}

\item{preprocess_data}{stored function that returns preprocessed data}
}
\value{
list object with user selected options and matrix build
}
\description{
This function sets up the Statmodel server to process data based on user
selected inputs
}
\examples{
NA

}
