% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-statmodel-server.R
\name{statmodelServer}
\alias{statmodelServer}
\title{Statmodel Server module for stat inference}
\usage{
statmodelServer(
  input,
  output,
  session,
  parent_session,
  loadpage_input,
  qc_input
)
}
\arguments{
\item{input}{input object to capture different ui element values}

\item{output}{to render and create elements}

\item{session}{session current module}

\item{parent_session}{session of the main calling module}

\item{loadpage_input}{input object from loadpage UI}

\item{qc_input}{input object from QC UI}
}
\value{
list object with user selected options and matrix build
}
\description{
This function sets up the Statmodel server to process data based on user
selected inputs
}
