% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_calculations.R
\name{tmt_model}
\alias{tmt_model}
\title{Main TMT modeling function for MSstatsShiny application}
\usage{
tmt_model(data, contrast.matrix, busy_indicator = TRUE)
}
\value{
list of TMT modeling results
}
\description{
Main TMT function to model MSstatsShiny data.
}
\examples{
#tmt_model(NULL)

}
