% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_calculations.R
\name{tmt_model}
\alias{tmt_model}
\title{Main TMT modeling function for MSstatsShiny application}
\usage{
tmt_model(data, input, contrast.matrix, busy_indicator = TRUE)
}
\arguments{
\item{data}{summarized data from output of MSstats summarization function.}

\item{input}{options for data processing input by the user}

\item{contrast.matrix}{contrast matrix specifying which conditions should be compared}

\item{busy_indicator}{Boolean indicator indicating whether or not to display
shiny waiting indicator.}
}
\value{
list of TMT modeling results
}
\description{
Main TMT function to model MSstatsShiny data.
}
\examples{
data("tmt_pd_model")
names(tmt_pd_model)
head(tmt_pd_model$ComparisonResult)

}
