% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_calculations.R
\name{tmt_summarization_loop}
\alias{tmt_summarization_loop}
\title{Main TMT summarization calculation function for MSstatsShiny application}
\usage{
tmt_summarization_loop(data, input, busy_indicator = TRUE)
}
\arguments{
\item{data}{Data converted into MSstats format.}

\item{input}{options for data processing input by the user}

\item{busy_indicator}{Boolean indicator indicating whether or not to display
shiny waiting indicator.}
}
\value{
list of TMT summarization results
}
\description{
Main TMT function to calculate MSstatsShiny results.
}
\examples{
data(raw.pd, package = "MSstatsTMT")
data(annotation.pd, package = "MSstatsTMT")

testdata <- MSstatsTMT::PDtoMSstatsTMTFormat(raw.pd, 
                                             annotation.pd,
                                             use_log_file = FALSE
                                             )

input = list()
input$summarization = "msstats"
input$norm = "equalizeMedians"
input$log = 2
input$names = NULL
input$features_used	= "all"
code_n_feat=3
input$censInt = "NA"
input$features_used	= "all"
input$MBi = TRUE
input$remove50 = FALSE
input$maxQC = 0.999
input$null = FALSE
input$null1 = FALSE
input$DDA_DIA = "LF"
input$global_norm = TRUE
input$reference_norm = TRUE
input$remove_norm_channel = TRUE
input$maxQC1 = NULL
input$moderated = FALSE

summarization_tmt_test = tmt_summarization_loop(testdata, input, 
                                               busy_indicator = FALSE)

}
