% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_calculations.R
\name{tmt_summarization_loop}
\alias{tmt_summarization_loop}
\title{Main TMT summarization calculation function for MSstatsShiny application}
\usage{
tmt_summarization_loop(data, busy_indicator = TRUE)
}
\value{
list of TMT summarization results
}
\description{
Main TMT function to calculate MSstatsShiny results.
}
\examples{
NULL
}
