% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-visualize-network-server.R
\name{visualizeNetworkServer}
\alias{visualizeNetworkServer}
\title{Server logic for network visualization module}
\usage{
visualizeNetworkServer(id, parent_session, dataComparison)
}
\arguments{
\item{id}{Module ID string}

\item{parent_session}{Parent Shiny session}

\item{dataComparison}{Reactive expression containing comparison data}
}
\value{
None (side effects only)
}
\description{
Server logic for network visualization module
}
